refactor: Order 로그인 권한 확인

+++ src/main/java/com/team2/fitinside/order/service/OrderAdminService.java
@@ -1,5 +1,9 @@
 package com.team2.fitinside.order.service;
 
+import com.team2.fitinside.config.SecurityUtil;
+import com.team2.fitinside.member.entity.Authority;
+import com.team2.fitinside.member.entity.Member;
+import com.team2.fitinside.member.repository.MemberRepository;
 import com.team2.fitinside.order.common.OrderStatus;
 import com.team2.fitinside.order.dto.OrderResponseDto;
 import com.team2.fitinside.order.dto.OrderStatusUpdateRequestDto;
@@ -11,6 +15,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.nio.file.AccessDeniedException;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -20,10 +25,11 @@ public class OrderAdminService {
 
     private final OrderMapper orderMapper;
     private final OrderRepository orderRepository;
+    private final MemberRepository memberRepository;
 
     // 전체 주문 조회
-    public List<OrderResponseDto> findAllOrdersByAdmin() {
-        // 전체 주문 조회 (isDeleted=false)
+    public List<OrderResponseDto> findAllOrdersByAdmin() throws AccessDeniedException {
+        checkAdmin();
         List<Order> orders = orderRepository.findAll().stream()
                 .filter(order -> !order.isDeleted())
                 .collect(Collectors.toList());
@@ -33,19 +39,30 @@ public class OrderAdminService {
 
     // 주문 상태 수정
     @Transactional
-    public OrderResponseDto updateOrderStatus(Long orderId, OrderStatusUpdateRequestDto request) throws Exception {
-        Order findOrder = orderRepository.findById(orderId).orElseThrow(() -> new OrderNotFoundException("주문이 존재하지 않습니다."));
+    public OrderResponseDto updateOrderStatus(Long orderId, OrderStatusUpdateRequestDto request) throws AccessDeniedException {
+        checkAdmin();
+        Order order = orderRepository.findById(orderId).orElseThrow(() -> new OrderNotFoundException("주문이 존재하지 않습니다."));
+
         OrderStatus status = OrderStatus.valueOf(request.getStatus().toUpperCase());
-        findOrder.updateOrderStatus(status);
-        Order updatedOrder = orderRepository.save(findOrder);
-        return orderMapper.toOrderResponseDto(updatedOrder);
+        order.updateOrderStatus(status);
+        return orderMapper.toOrderResponseDto(order);
     }
 
     // 주문 삭제
     @Transactional
-    public void deleteOrder(Long orderId) throws Exception {
+    public void deleteOrder(Long orderId) throws AccessDeniedException {
+        checkAdmin();
         Order findOrder = orderRepository.findById(orderId).orElseThrow(() -> new OrderNotFoundException("주문이 존재하지 않습니다."));
         findOrder.deleteOrder();
+    }
+
+    private void checkAdmin() throws AccessDeniedException {
+        Long loginMemberId = SecurityUtil.getCurrentMemberId();
+        Member findMember = memberRepository.findById(loginMemberId).orElseThrow(() -> new IllegalStateException("인증된 사용자의 정보가 유효하지 않습니다."));
+
+        if (findMember.getAuthority() != Authority.ROLE_ADMIN) {
+            throw new AccessDeniedException("관리자 권한이 필요합니다.");
+        }
     }
 
 }
